refactor(admin): extract MonthView and EventSummary components
